Revert 1606 to allow routing of /index.php/foo again
PR #1606 changed the way that the Uri's baseUrl was detected in order to
fix #1590. However, this change meant that routing of URLs of the form
/foo/index.php/bar/baz no longer worked.

This commit reverts #1606 so that routing works again as the stripping
off of the script name  when using rewriting could be done at the web
server configuration level, or via a route the redirects to the one
without the script name in it.